Update toStringArray divide
